pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
